# public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Pede música API doc"
  },
  "host": "ask-for-music.vercel.app",
  "basePath": "/api",
  "schemes": ["https", "http"],
  "paths": {
    "/verify-auth": {
      "post": {
        "summary": "Verify musician auth code",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Auth"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success verification"
          },
          "401": {
            "description": "Invalid auth code"
          }
        }
      }
    },

    "/songs": {
      "post": {
        "summary": "Musician add a knowed song",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Song"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success verification",
            "schema": {
              "$ref": "#/definitions/IdBody"
            }
          },
          "422": {
            "description": "Missing parameters!"
          },
          "500": {
            "description": "Failed request"
          }
        }
      },

      "get": {
        "summary": "Get all song knowed by the musician",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "An array of songs",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Song"
              }
            }
          },
          "500": {
            "description": "Failed request"
          }
        }
      },

      "delete": {
        "summary": "Delete a song by ID",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/IdBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success deleted"
          },
          "422": {
            "description": "Missing parameters!"
          },
          "500": {
            "description": "Failed request"
          }
        }
      }
    },

    "/requests": {
      "post": {
        "summary": "Send a song request to musician",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SongRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success verification",
            "schema": {
              "$ref": "#/definitions/IdBody"
            }
          },
          "422": {
            "description": "Missing parameters!"
          },
          "500": {
            "description": "Failed request"
          }
        }
      },

      "get": {
        "summary": "Get all song requests",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "An array of song requests",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SongRequest"
              }
            }
          },
          "500": {
            "description": "Failed request"
          }
        }
      },

      "delete": {
        "summary": "Delete a song request by ID",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/IdBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success deleted"
          },
          "422": {
            "description": "Missing parameters!"
          },
          "500": {
            "description": "Failed request"
          }
        }
      }
    }
  },
  "definitions": {
    "SongRequest": {
      "type": "object",
      "required": ["requestedBy", "songId"],
      "properties": {
        "requestedBy": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "songId": {
          "type": "string"
        }
      }
    },
    "Song": {
      "type": "object",
      "required": ["name", "genre"],
      "properties": {
        "artist": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "genre": {
          "type": "string"
        }
      }
    },
    "Auth": {
      "type": "object",
      "required": ["code"],
      "properties": {
        "code": {
          "type": "string"
        }
      }
    },
    "IdBody": {
      "type": "object",
      "required": ["id"],
      "properties": {
        "id": {
          "type": "string"
        }
      }
    }
  }
}